refactor: create database folder
